+++ docs/CHANGELOG.md
@@ -1,4 +1,3 @@
-
 # Changelog
 
 ## Unreleased
@@ -8,6 +7,8 @@
 - Dictionary build `--no-output` option
 - `DictionaryOutputOptions` class for `DictionaryBuilder.build`
 - `BuildSummary` class for results of `DictionaryBuilder.build`
+- Experimental pipelines for dictionary builds
+- `ignore_case` option to `GrasciiValidator`
 
 ### Changed